+++ Android/app/src/main/java/tie/hackathon/travelguide/MainActivity.java
@@ -16,17 +16,13 @@ import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
-import android.util.Log;
 import android.view.MenuItem;
 
-import com.estimote.sdk.Beacon;
 import com.estimote.sdk.BeaconManager;
 import com.estimote.sdk.Region;
 
-import java.util.List;
-import java.util.UUID;
-
 import Util.Constants;
+import tie.hackathon.travelguide.login.LoginActivity;
 
 /**
  * Launcher Activity; Handles fragment changes;
